Update doc gen to work on windows
No longer relying on groff nor col to generate
output.  It's not quite as nice as the manpage output
yet, primarily with how it handles indentation, but
all the heavy lifting is done in botocore's text writer
with some small modifications.

I also took this opportunity to increase the test coverage
of the doc index.  Because mose of this deals with integrating
with AWS CLI interfaces, I added a few integration tests for
doc indexing.