Moved CompletionIndex.load_completions call from init.
Moving this out is more modular, as CompletionIndex() is created in two places in the source: main and lexer (possible refactor TODO item).  It's also called once in the tests.  This change also simplifies testing as we can more easily substitute a mock function, although this could also probably be achieved with a mock side effect.